Fix coloring
In 11.0.0, Blizzard has changed (maybe by mistake, since the documentation doesn't mention it) the color you get from the GetBreedQuality function.
Now it is indexed from 0, so subtracting 1 breaks everything.
It would make sense for it to be on purpose, since everything else related to those colors is indexed from 0.